Defined hazard
Hazard is now implemented as a set of rates, each with a min and max number. When a number is randomly generated between the min and max, it is used to determine how many rate elements are activated as hazard. For instance, a single HazardList would contain 3 hazard events, and if the random  number generated was 2, 2 of the hazard events would activate, but the 3rd wouldn't